criado metodo para formatar dinheiro. listando origem e valor.

+++ app/src/main/java/com/blimas/desafiotimelineandroid/service/model/LancamentoModel.kt
@@ -8,7 +8,7 @@ class LancamentoModel {
     var id: Int = 0
 
     @SerializedName("valor")
-    var valor: Float = 0.0F
+    var valor: Double = 0.0
 
     @SerializedName("origem")
     var origem: String = ""
